Refactor ContentVuer to use dynamic components for each viewers.
